refactor(agent,channels): split loop and startup support
