+++ src/agent/mod.rs
@@ -6,6 +6,7 @@ pub(crate) mod history;
 pub mod loop_;
 pub mod memory_loader;
 pub mod prompt;
+pub(crate) mod runtime_support;
 pub(crate) mod tool_call_parser;
 pub(crate) mod tool_filter;
